Turn incorrect panic into error
This line:

--- something

was considered a meta separator when we peeked into first 4 characters
(with trimmed space), but then when we read line, it turned out that
it was not the correct separator (which should be on its own line). We
considered this a programmer error and panicked, but it's actually a
normal error.