Add sitemap generation
Follows https://sitemaps.org spec.

If site.yml has a sitemap filename, for example:

	sitemap: /sitemap.xml

it will be rendered into this location in out directory.

Sitemap parameters are controlled from meta:

	changefreq: weekly
	priority: 0.8

A file can be excluded from the site map using:

	sitemap: false

in the meta.

The last change date is taken from the file modification time.

Only pages that are rendered with .htm, .html, and .xml extension are
included in the map (in addition to posts and tag pages).